# zstd: Speed up + improve best encoder